+++ module/exporter-file/src/Application/Model/ExporterFileConfigurationModel.php
@@ -8,9 +8,13 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterFile\Application\Model;
 
+use Ergonode\ExporterFile\Application\Validator\SegmentNotChanged;
 use Symfony\Component\Validator\Constraints as Assert;
 use Ergonode\ExporterFile\Domain\Entity\FileExportChannel;
 
+/**
+ * @SegmentNotChanged(groups={"Update"})
+ */
 class ExporterFileConfigurationModel
 {
     /**
@@ -38,13 +42,16 @@ class ExporterFileConfigurationModel
 
     public ?string $segmentId = null;
 
+    public ?FileExportChannel $channel = null;
+
     public function __construct(FileExportChannel $channel = null)
     {
         if ($channel) {
+            $this->channel = $channel;
             $this->name = $channel->getName();
             $this->format = $channel->getFormat();
             $this->exportType = $channel->getExportType();
-            $this->segmentId =  $channel->getSegmentId() ? $channel->getSegmentId()->getValue() : null;
+            $this->segmentId = $channel->getSegmentId() ? $channel->getSegmentId()->getValue() : null;
             $this->languages = $channel->getLanguages();
         }
     }