+++ module/segment/src/Infrastructure/Persistence/Query/DbalSegmentProductsQuery.php
@@ -11,7 +11,6 @@ namespace Ergonode\Segment\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
-use Doctrine\DBAL\Types\Types;
 use Ergonode\Segment\Domain\Query\SegmentProductsQueryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\SegmentId;
 
@@ -19,7 +18,6 @@ class DbalSegmentProductsQuery implements SegmentProductsQueryInterface
 {
     private const PRODUCT_TABLE = 'public.product';
     private const SEGMENT_PRODUCT_TABLE = 'public.segment_product';
-    private const AUDIT_TABLE = 'audit';
 
     private Connection $connection;
 
@@ -42,37 +40,6 @@ class DbalSegmentProductsQuery implements SegmentProductsQueryInterface
             ->setParameter(':segmentId', $segmentId->getValue())
             ->execute()
             ->fetchAll(\PDO::FETCH_COLUMN);
-    }
-
-    /**
-     * @return array
-     */
-    public function getAllEditedProducts(SegmentId $segmentId, ?\DateTime $dateTime = null): array
-    {
-        $qb = $this->getQuery();
-
-        $qb->select('sp.product_id')
-            ->join('sp', self::AUDIT_TABLE, 'a', 'a.id = sp.product_id')
-            ->where($qb->expr()->eq('segment_id', ':segmentId'))
-            ->andWhere($qb->expr()->eq('available', ':available'))
-            ->setParameter(':available', true)
-            ->setParameter(':segmentId', $segmentId->getValue());
-
-        if ($dateTime) {
-            $qb
-                ->andWhere($qb->expr()->gte('edited_at', ':editedAt'))
-                ->setParameter(':editedAt', $dateTime, Types::DATETIMETZ_MUTABLE);
-        }
-
-        $result = $qb
-            ->execute()
-            ->fetchAll(\PDO::FETCH_COLUMN);
-
-        if (false !== $result) {
-            return $result;
-        }
-
-        return [];
     }
 
     /**